+++ CHANGELOG.md
@@ -1,3 +1,6 @@
-## 1.0.0
+## 1.0.1
+- Modify display screen of `Windows11`.
+- Update README file and API documents.
 
-* Implements UI for `Windows10` and `Windows11` systems with support English only.
+## 1.0.0
+- Initial release with support English only.